backend/requirements.txt
fastapi
uvicorn
sqlalchemy
loguru
celery
redis
psycopg2-binary
python-dotenv
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
